move URLListField to its own file
